Let to_rgba return uint8; track changes to cmap
Images require rgba as 4 uint8s, so it is more efficient
to generate these directly in to_rgba than to generate 4
doubles and convert them later.

The tracking of changes in ScalarMappable was handling
communication between objects, but was not keeping track of
when to_rgba needs to be rerun.  A dictionary was added
to do this.


git-svn-id: https://matplotlib.svn.sourceforge.net/svnroot/matplotlib@4374 f61c4167-ca0d-0410-bb4a-bb21726e55ed