Experimental clipping of Line _transformed_path to speed zoom and pan.
This can be modified to work with x monotonically decreasing, but
for a first try it works only with x monotonically increasing.
The intention is to greatly speed up zooming and panning into
a small segment of a very long time series (e.g., 10^6 points)
without incurring any significant speed penalty in other situations.


git-svn-id: https://matplotlib.svn.sourceforge.net/svnroot/matplotlib@7100 f61c4167-ca0d-0410-bb4a-bb21726e55ed