make default argument to Axes.get_xaxis_transform() not raise error
I made this a separate commit from the previous one because it may be
useful to see the errors that arise if get_xaxis_transform() or
get_yaxis_transform() are called with no arguments. This would allow
one to figure out which code paths are calling these methods and fix
them if needed rather than silently partially succeeding.

git-svn-id: https://matplotlib.svn.sourceforge.net/svnroot/matplotlib@7145 f61c4167-ca0d-0410-bb4a-bb21726e55ed