alpha is left as None until you really want to specify it; closes 2957321.
I think this, together with some other recent commits, fixes various
longstanding problems in the way alpha has been handled.  The default
value is left as None, and with this value, existing alpha specified
in RGBA sequences will not be overridden.  Substitution of the default
value of 1 for a value of None is delayed until the GraphicsContextBase
set_alpha method is executed.



git-svn-id: https://matplotlib.svn.sourceforge.net/svnroot/matplotlib@8372 f61c4167-ca0d-0410-bb4a-bb21726e55ed